Create sourceCode.md
This is for my girls Who Code Summer Immersion Program TA assignment. It generates a random number between 1 and 3, which is associated with the terms "Rock, Paper, or Scissors". It compares it to the user input and determines a winner, looping through as many rematches as the user wants to play.